chore(recoil): move route
